+++ CRM/Civirules/Form/Rule.php
@@ -125,7 +125,10 @@ class CRM_Civirules_Form_Rule extends CRM_Core_Form {
     $session = CRM_Core_Session::singleton();
     $userId = $session->get('userID');
 
-    if (isset($this->_submitValues['_qf_Rule_next_clone'])) {
+    if (isset($this->_submitValues['_qf_Rule_next_cancel'])) {
+      CRM_Utils_System::redirect(CRM_Utils_System::url('civicrm/civirules/form/rulesview', 'reset=1'));
+    }
+    elseif (isset($this->_submitValues['_qf_Rule_next_clone'])) {
       $result = civicrm_api3('CiviRuleRule', 'clone', [
         'id' => $this->ruleId,
       ]);
@@ -134,7 +137,10 @@ class CRM_Civirules_Form_Rule extends CRM_Core_Form {
     } else {
       $this->saveRule($this->_submitValues, $userId);
       $this->saveRuleTrigger($this->_submitValues);
-      $session->setStatus('Rule with linked Trigger saved succesfully', 'CiviRule saved', 'success');
+      $session->setStatus("Rule: '{$this->_submitValues['rule_label']}' saved succesfully", 'CiviRule saved', 'success');
+      if (isset($this->_submitValues['_qf_Rule_upload_done'])) {
+        CRM_Utils_System::redirect(CRM_Utils_System::url('civicrm/civirules/form/rulesview', 'reset=1'));
+      }
     }
     /*
      * if add mode, set user context to form in edit mode to add conditions and actions
@@ -274,8 +280,13 @@ class CRM_Civirules_Form_Rule extends CRM_Core_Form {
     } else {
       $this->addButtons(array(
         array('type' => 'next', 'name' => ts('Save'), 'isDefault' => TRUE,),
+        [
+          'type' => 'upload',
+          'name' => ts('Save and Done'),
+          'subName' => 'done',
+        ],
         array('type' => 'next', 'name' => ts('Clone'), 'subName' => 'clone', 'icon' => 'fa-creative-commons'),
-        array('type' => 'cancel', 'name' => ts('Cancel'))));
+        array('type' => 'next', 'name' => ts('Close'), 'subName' => 'cancel', 'icon' => 'fa-close')));
     }
   }