+++ CRM/CivirulesConditions/Case/StatusChanged.php
@@ -3,60 +3,21 @@
 class CRM_CivirulesConditions_Case_StatusChanged extends CRM_CivirulesConditions_Generic_FieldValueChangeComparison {
 
   /**
-   * Returns the value of the field for the condition
-   * For example: I want to check if age > 50, this function would return the 50
+   * Returns name of entity
    *
-   * @param object CRM_Civirules_TriggerData_TriggerData $triggerData
-   * @return mixed
-   * @access protected
+   * @return string
    */
-  protected function getOriginalFieldValue(CRM_Civirules_TriggerData_TriggerData $triggerData) {
-    $field = 'status_id';
-
-    $data = $triggerData->getOriginalData();
-    if (isset($data[$field])) {
-      return $data[$field];
-    }
-
-    return null;
+  protected function getEntity() {
+    return 'Case';
   }
 
   /**
-   * Returns the value of the field for the condition
-   * For example: I want to check if age > 50, this function would return the 50
+   * Returns name of the field
    *
-   * @param object CRM_Civirules_TriggerData_TriggerData $triggerData
-   * @return mixed
-   * @access protected
+   * @return string
    */
-  protected function getFieldValue(CRM_Civirules_TriggerData_TriggerData $triggerData) {
-    $field = 'status_id';
-
-    $data = $triggerData->getEntityData('Case');
-    if (isset($data[$field])) {
-      return $data[$field];
-    }
-
-    return null;
-  }
-
-  /**
-   * This function validates whether this condition works with the selected trigger.
-   *
-   * This function could be overriden in child classes to provide additional validation
-   * whether a condition is possible in the current setup. E.g. we could have a condition
-   * which works on contribution or on contributionRecur then this function could do
-   * this kind of validation and return false/true
-   *
-   * @param CRM_Civirules_Trigger $trigger
-   * @param CRM_Civirules_BAO_Rule $rule
-   * @return bool
-   */
-  public function doesWorkWithTrigger(CRM_Civirules_Trigger $trigger, CRM_Civirules_BAO_Rule $rule) {
-    if ($trigger instanceof CRM_Civirules_TriggerData_Interface_OriginalData) {
-      return $trigger->doesProvideEntity('Case');
-    }
-    return false;
+  protected function getEntityStatusFieldName() {
+    return 'status_id';
   }
 
   /**
@@ -64,14 +25,14 @@ class CRM_CivirulesConditions_Case_StatusChanged extends CRM_CivirulesConditions
    * case the field is a select field, e.g. gender, or financial type
    * Return false when the field is a select field
    *
-   * This method could be overriden by child classes to return the option
+   * This method could be overridden by child classes to return the option
    *
    * The return is an array with the field option value as key and the option label as value
    *
-   * @return bool
+   * @return array
    */
   public function getFieldOptions() {
-    return CRM_Core_BAO_OptionValue::getOptionValuesAssocArrayFromName('case_status');
+    return CRM_CivirulesConditions_Case_CaseStatus::getEntityStatusList(TRUE);
   }
 
   /**