+++ CRM/CivirulesConditions/Contribution/Status.php
@@ -1,75 +1,32 @@
 <?php
 
-class CRM_CivirulesConditions_Contribution_Status extends CRM_Civirules_Condition {
-
-  private $conditionParams = array();
+class CRM_CivirulesConditions_Contribution_Status extends CRM_CivirulesConditions_Generic_Status {
 
   /**
-   * Method to set the Rule Condition data
-   *
-   * @param array $ruleCondition
-   * @access public
-   */
-  public function setRuleConditionData($ruleCondition) {
-    parent::setRuleConditionData($ruleCondition);
-    $this->conditionParams = array();
-    if (!empty($this->ruleCondition['condition_params'])) {
-      $this->conditionParams = unserialize($this->ruleCondition['condition_params']);
-    }
-  }
-
-  public function isConditionValid(CRM_Civirules_TriggerData_TriggerData $triggerData) {
-    $contribution = $triggerData->getEntityData('Contribution');
-    if ($contribution['contribution_status_id'] == $this->conditionParams['contribution_status_id']) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Returns a redirect url to extra data input from the user after adding a condition
-   *
-   * Return false if you do not need extra data input
-   *
-   * @param int $ruleConditionId
-   * @return bool|string
-   * @access public
-   * @abstract
-   */
-  public function getExtraDataInputUrl($ruleConditionId) {
-    return CRM_Utils_System::url('civicrm/civirule/form/condition/contributionstatus/', 'rule_condition_id='.$ruleConditionId);
-  }
-
-  /**
-   * Returns a user friendly text explaining the condition params
-   * e.g. 'Older than 65'
-   *
+   * The entity name (eg. Membership)
    * @return string
-   * @access public
-   * @throws \CiviCRM_API3_Exception
    */
-  public function userFriendlyConditionParams() {
-    $contributionStatus =  $status = civicrm_api3('OptionValue', 'getvalue', array(
-      'return' => 'label',
-      'option_group_id' => 'contribution_status',
-      'value' => $this->conditionParams['contribution_status_id']));
-    return 'Contribution status is '.$contributionStatus;
+  protected function getEntity() {
+    return 'Contribution';
   }
 
   /**
-   * This function validates whether this condition works with the selected trigger.
-   *
-   * This function could be overriden in child classes to provide additional validation
-   * whether a condition is possible in the current setup. E.g. we could have a condition
-   * which works on contribution or on contributionRecur then this function could do
-   * this kind of validation and return false/true
-   *
-   * @param CRM_Civirules_Trigger $trigger
-   * @param CRM_Civirules_BAO_Rule $rule
-   * @return bool
+   * The entity status field (eg. membership_status_id)
+   * @return string
    */
-  public function doesWorkWithTrigger(CRM_Civirules_Trigger $trigger, CRM_Civirules_BAO_Rule $rule) {
-    return $trigger->doesProvideEntity('Contribution');
+  public function getEntityStatusFieldName() {
+    return 'contribution_status_id';
+  }
+
+  /**
+   * Returns an array of statuses as [ id => label ]
+   * @param bool $active
+   * @param bool $inactive
+   *
+   * @return array
+   */
+  public static function getEntityStatusList($active = TRUE, $inactive = FALSE) {
+    return parent::getEntityStatusListFromOptionGroup('contribution_status', $active, $inactive);
   }
 
 }