+++ CRM/CivirulesConditions/Form/ContributionRecur/PaymentProcessor.php
@@ -71,10 +71,16 @@ class CRM_CivirulesConditions_Form_ContributionRecur_PaymentProcessor extends CR
     parent::postProcess();
   }
 
+  /**
+   * Returns a help text for this condition.
+   * The help text is shown to the administrator who is configuring the condition.
+   *
+   * @return string
+   */
   protected function getHelpText() {
-    $help = E::ts('This condition works with Membership (checks the linked recur) and Recurring Contribution triggers.');
-    $help .= '<ul><li>' . E::ts('IS ONE OF will match a Membership/Recurring Contribution that has a payment processor in the list.');
-    $help .= '</li><li>' . E::ts('IS NOT ONE OF will match a Membership that does not have a linked recurring contribution OR a Membership/Recurring Contribution that: Has a recurring contribution with no payment processor; Has a recurring contribution that has a payment processor not in the list.');
+    $help = E::ts('This condition works with Membership (checks the linked recur), Contribution (checks the linked recur) and Recurring Contribution triggers.');
+    $help .= '<ul><li>' . E::ts('IS ONE OF will match a Membership/Contribution/Recurring Contribution that has a payment processor in the list.');
+    $help .= '</li><li>' . E::ts('IS NOT ONE OF will match a Membership/Contribution that does not have a linked recurring contribution OR a Membership/Contribution/Recurring Contribution that: Has a recurring contribution with no payment processor; Has a recurring contribution that has a payment processor not in the list.');
     $help .= '</ul>';
     return $help;
   }