+++ CRM/CivirulesCronTrigger/MembershipEndDate.php
@@ -4,24 +4,28 @@
  * @license http://www.gnu.org/licenses/agpl-3.0.html
  */
 
+use CRM_Civirules_ExtensionUtil as E;
+
 class CRM_CivirulesCronTrigger_MembershipEndDate extends CRM_Civirules_Trigger_Cron {
 
-  private $dao = false;
+  /**
+   * @var \CRM_Member_DAO_Membership $dao
+   */
+  private $dao = NULL;
 
   public static function intervals() {
-    return array(
+    return [
       '-days' => ts('Day(s) before end date'),
       '-weeks' => ts('Week(s) before end date'),
       '-months' => ts('Month(s) before end date'),
       '+days' => ts('Day(s) after end date'),
       '+weeks' => ts('Week(s) after end date'),
       '+months' => ts('Month(s) after end date'),
-    );
+    ];
   }
 
   /**
    * This function returns a CRM_Civirules_TriggerData_TriggerData this entity is used for triggering the rule
-   *
    * Return false when no next entity is available
    *
    * @return CRM_Civirules_TriggerData_TriggerData|false
@@ -33,12 +37,23 @@ class CRM_CivirulesCronTrigger_MembershipEndDate extends CRM_Civirules_Trigger_C
       }
     }
     if ($this->dao->fetch()) {
-      $data = array();
+      $data = [];
       CRM_Core_DAO::storeValues($this->dao, $data);
       $triggerData = new CRM_Civirules_TriggerData_Cron($this->dao->contact_id, 'Membership', $data);
+
+      if ($this->dao->contribution_recur_id) {
+        $contributionRecur = new CRM_Contribute_BAO_ContributionRecur();
+        $contributionRecur->id = $this->dao->contribution_recur_id;
+        if ($contributionRecur->find(TRUE)) {
+          $data = array();
+          CRM_Core_DAO::storeValues($contributionRecur, $data);
+          $triggerData->setEntityData('ContributionRecur', $data);
+        }
+      }
+
       return $triggerData;
     }
-    return false;
+    return FALSE;
   }
 
   /**
@@ -52,53 +67,58 @@ class CRM_CivirulesCronTrigger_MembershipEndDate extends CRM_Civirules_Trigger_C
 
   /**
    * Method to query trigger entities
-   *
-   * @access private
    */
   private function queryForTriggerEntities() {
     if (empty($this->triggerParams['membership_type_id'])) {
       return false;
     }
 
-    $params[1] = array($this->triggerParams['membership_type_id'], 'Integer');
+    // membership_type_id used to be a single value, but now we can have multiple membership types
+    if (is_array($this->triggerParams['membership_type_id'])) {
+      $params[1] = [implode(',', $this->triggerParams['membership_type_id']), 'CommaSeparatedIntegers'];
+    }
+    else {
+      $params[1] = [$this->triggerParams['membership_type_id'], 'Integer'];
+    }
+
     $end_date_statement = "AND DATE(m.end_date) = CURRENT_DATE()";
     switch ($this->triggerParams['interval_unit']) {
       case '-days':
         $end_date_statement = "AND DATE_SUB(m.end_date, INTERVAL %2 DAY) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
       case '-weeks':
         $end_date_statement = "AND DATE_SUB(m.end_date, INTERVAL %2 WEEK) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
       case '-months':
         $end_date_statement = "AND DATE_SUB(m.end_date, INTERVAL %2 MONTH) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
       case '+days':
         $end_date_statement = "AND DATE_ADD(m.end_date, INTERVAL %2 DAY) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
       case '+weeks':
         $end_date_statement = "AND DATE_ADD(m.end_date, INTERVAL %2 WEEK) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
       case '+months':
         $end_date_statement = "AND DATE_ADD(m.end_date, INTERVAL %2 MONTH) = CURRENT_DATE()";
-        $params[2] = array($this->triggerParams['interval'], 'Integer');
+        $params[2] = [$this->triggerParams['interval'], 'Integer'];
         break;
     }
 
     $sql = "SELECT m.*
             FROM `civicrm_membership` `m`
-            WHERE `m`.`membership_type_id` = %1
-            {$end_date_statement} 
+            WHERE `m`.`membership_type_id` IN (%1)
+            {$end_date_statement}
             AND `m`.`contact_id` NOT IN (
               SELECT `rule_log`.`contact_id`
               FROM `civirule_rule_log` `rule_log`
               WHERE `rule_log`.`rule_id` = %3 AND DATE(`rule_log`.`log_date`) = DATE(NOW())
             )";
-    $params[3] = array($this->ruleId, 'Integer');
+    $params[3] = [$this->ruleId, 'Integer'];
     $this->dao = CRM_Core_DAO::executeQuery($sql, $params, true, 'CRM_Member_DAO_Membership');
 
     return true;
@@ -111,13 +131,14 @@ class CRM_CivirulesCronTrigger_MembershipEndDate extends CRM_Civirules_Trigger_C
    *
    * @param int $ruleId
    * @return bool|string
-   * @access public
-   * @abstract
    */
   public function getExtraDataInputUrl($ruleId) {
     return CRM_Utils_System::url('civicrm/civirule/form/trigger/membershipenddate/', 'rule_id='.$ruleId);
   }
 
+  /**
+   * @param string $triggerParams
+   */
   public function setTriggerParams($triggerParams) {
     $this->triggerParams = unserialize($triggerParams);
   }
@@ -126,18 +147,39 @@ class CRM_CivirulesCronTrigger_MembershipEndDate extends CRM_Civirules_Trigger_C
    * Returns a description of this trigger
    *
    * @return string
-   * @access public
-   * @abstract
    */
   public function getTriggerDescription() {
-    $membership_types = CRM_Civirules_Utils::getMembershipTypes();
-    $interval_units = self::intervals();
+    $membershipTypes = CRM_Civirules_Utils::getMembershipTypes();
+    $intervalUnits = self::intervals();
+    $intervalUnitLabel = $intervalUnits[$this->triggerParams['interval_unit']];
 
+    if (is_array($this->triggerParams['membership_type_id'])) {
+      $membershipTypeLabels = [];
+      foreach ($this->triggerParams['membership_type_id'] as $membershipTypeID) {
+        $membershipTypeLabels[] = $membershipTypes[$membershipTypeID];
+      }
+      $membershipTypeLabel = implode(',', $membershipTypeLabels);
+    }
+    else {
+      $membershipTypeLabel = $membershipTypes[$this->triggerParams['membership_type_id']];
+    }
 
-    $membershipTypeLabel = $membership_types[$this->triggerParams['membership_type_id']];
-    $intervalUnitLabel = $interval_units[$this->triggerParams['interval_unit']];
+    return E::ts('Membership Types %1 - %2 %3', [
+      1 => $membershipTypeLabel,
+      2 => $this->triggerParams['interval'],
+      3 => $intervalUnitLabel,
+    ]);
+  }
 
-    return ts('Membership end date with type %1 %2 %3', array(1 => $membershipTypeLabel, 2=> $this->triggerParams['interval'], $intervalUnitLabel));
+  /**
+   * Returns additional entities provided in this trigger.
+   *
+   * @return array of CRM_Civirules_TriggerData_EntityDefinition
+   */
+  protected function getAdditionalEntities() {
+    $entities = parent::getAdditionalEntities();
+    $entities[] = new CRM_Civirules_TriggerData_EntityDefinition('ContributionRecur', 'ContributionRecur', 'CRM_Contribute_DAO_ContributionRecur' , 'ContributionRecur');
+    return $entities;
   }
 
 }